Fix gc stats race condition and add timestamp of the gc run
Previously, our counters on the number of GC runs were being read from a
static struct when constructing the stats event to emit.  In blocks of
synchronous code, GC can run multiple times before control returns to
the event loop where the stats events will be constructed and emitted.
All events would then get the same values for the counters at the time
the events are created.  This passes a copy of the counters with the
baton to fix this.

This also adds a us timestamp of when the gc run occurred, which can be
useful for sequencing events.